Link legacy and ORD output executables
